Fix alert entity id type to match schema (#20)

+++ src/main/java/com/investment/metal/database/AlertRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AlertRepository extends JpaRepository<Alert, Long> {
+public interface AlertRepository extends JpaRepository<Alert, Integer> {
 
     Optional<List<Alert>> findByUserId(long userId);
